requirements.txt
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytube==15.0.0
pytz==2025.2
pyzmq==26.3.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.2
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
yt-dlp==2025.5.22
